+++ src/net/ssehub/kernel_haven/metric_haven/metric_components/weights/IVariableWeight.java
@@ -18,7 +18,7 @@ public interface IVariableWeight {
      * @param variable An existing variable of the variability model, for which the weight shall be returned for.
      * @return A positive weight (&ge; 1), 1 denotes a neutral value.
      */
-    public int getWeight(String variable);
+    public long getWeight(String variable);
     
     /**
      * Returns the weight for the specified variable and the currently measured code artifact.
@@ -26,7 +26,7 @@ public interface IVariableWeight {
      * @param codefile The currently measured code artifact (only required for some weights).
      * @return A positive weight (&ge; 1), 1 denotes a neutral value.
      */
-    public default int getWeight(String variable, File codefile) {
+    public default long getWeight(String variable, File codefile) {
         return getWeight(variable);
     }
